+++ CHANGELOG.md
@@ -9,11 +9,16 @@
 ## 0.0.3
 
 * Added support to get device details, app details 
-  and to read launch data (to read data from
-  intent(android) & urlScheme(iOS)).
+  and to read launch data to read data from
+  intent(android) & urlScheme (iOS).
 
 * Added params support for launchApp function.
 
 * Improved documentation & updated README file.
 
 * Fixed bugs.
+
+## 0.0.4
+
+* Added support to open the device settings page.
+